Move ContactRegistration to v1 in Telemetry
