+++ EventsHandler/Api/EventsHandler/Properties/Resources.resx
@@ -368,6 +368,9 @@
   <data name="Configuration_ERROR_OpenZaakVersionUnknown" xml:space="preserve">
     <value>The version of OpenZaak service to be used by OMC is unknown or not supported ('Features:OpenServicesVersion').</value>
   </data>
+  <data name="Configuration_ERROR_TelemetryVersionUnknown" xml:space="preserve">
+    <value>The version of Telemetry service to be used by OMC is unknown or not supported ('Features:OpenServicesVersion').</value>
+  </data>
   <data name="Events_NotifyClientInitialized" xml:space="preserve">
     <value>The Notify HttpClient was initialized for organization ID:</value>
   </data>
